requirements.txt
Django==5.0.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.15.2
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7